Create 9. Odd or Even
we used two methods
method-1 
normal way
if number%2 is 0 it is even otherwise odd
because all even numbers are divisible by 2 
method -2
Optimal Solution :
Algorithm:
The oddEven method in your Java class determines whether a given integer N is odd or even using bitwise AND operation with 1. Here's the algorithm:
Input: Take an integer N as input.
Check Odd/Even:
Use the bitwise AND operation (&) between N and 1.
If the result is 0, then N is even because an even number's least significant bit is 0. Return "even".
Otherwise, if the result is 1, then N is odd because an odd number's least significant bit is 1. Return "odd".